+++ app/filters/dorsale/customer_vault/small_data/filter_for_events.rb
@@ -1,0 +1,5 @@
+class Dorsale::CustomerVault::SmallData::FilterForEvents < ::Agilibox::SmallData::Filter
+  STRATEGIES = {
+    "event_action" => ::Agilibox::SmallData::FilterStrategyByKeyValues.new(:action),
+  }
+end