+++ app/models/dorsale/customer_vault/event.rb
@@ -7,9 +7,8 @@ class Dorsale::CustomerVault::Event < ::Dorsale::ApplicationRecord
   belongs_to :person,  class_name: Dorsale::CustomerVault::Person
   belongs_to :comment, class_name: Dorsale::Comment
 
-  validates :author,  presence: true
   validates :person,  presence: true
-  validates :action,  presence: true, inclusion: {in: ACTIONS}
+  validates :action,  presence: true, inclusion: {in: proc {ACTIONS} }
   validates :comment, presence: true, if: proc { action == "comment" }
 
   default_scope -> {
@@ -20,9 +19,5 @@ class Dorsale::CustomerVault::Event < ::Dorsale::ApplicationRecord
 
   def date
     created_at.try(:to_date)
-  end
-
-  def text
-    t("text.#{action}")
   end
 end