+++ spec/controllers/dorsale/comments_controller_spec.rb
@@ -60,5 +60,18 @@ describe Dorsale::CommentsController, type: :controller do
       delete :destroy, params: {id: comment, back_url: "/"}
       expect(response).to redirect_to("/")
     end
+
+    describe "on person" do
+      it "should delete event" do
+        person  = create(:customer_vault_corporation)
+        comment = create(:dorsale_comment, commentable: person)
+        event   = create(:customer_vault_event, person: person, comment: comment, action: "comment")
+
+        delete :destroy, params: {id: comment}
+
+        expect { comment.reload }.to raise_error(ActiveRecord::RecordNotFound)
+        expect { event.reload   }.to raise_error(ActiveRecord::RecordNotFound)
+      end
+    end # describe "on person"
   end
 end