+++ spec/controllers/dorsale/customer_vault/people_controller_spec.rb
@@ -77,7 +77,7 @@ RSpec.describe ::Dorsale::CustomerVault::PeopleController, type: :controller do
 
         get :index
 
-        expect(assigns(:people)).to eq [corpo1, individual]
+        expect(assigns(:people)).to contain_exactly(corpo1, individual)
       end
     end # describe "filters"
 
@@ -91,20 +91,6 @@ RSpec.describe ::Dorsale::CustomerVault::PeopleController, type: :controller do
       end
     end # describe "search"
   end # describe "#list"
-
-  describe "#activity" do
-    before do
-      @corporation_1 = create(:customer_vault_corporation)
-      @corporation_2 = create(:customer_vault_corporation)
-      @event_1       = create(:customer_vault_event, person: @corporation_1, created_at: "2012-02-15")
-      @event_2       = create(:customer_vault_event, person: @corporation_2, created_at: "2012-03-15")
-    end
-
-    it "should assigns all events ordered by created_at DESC" do
-      get :activity
-      expect(assigns(:events)).to eq [@event_2, @event_1]
-    end
-  end # describe "#activity"
 
   describe "#create" do
     before do