+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Next version
 
 - Change nilify_blanks config
+- CustomerVault Events : persist text
+- CustomerVault : delete event when deleting comment
+- Comments : fix double comments JS
+- CustomerVault : do not validate events author
+- CustomerVault : move PeopleController#activity to EventsController#index
+- CustomerVault : add events filters
 
 ## 3.7.3
 - Fix copy action